Refactor GCN based on graphsage
